fix(api): Added map_to_docket_entry_sorting to the V3 RECAP search type
