+++ cl/lib/search_utils.py
@@ -23,6 +23,7 @@ from cl.lib.utils import (
     cleanup_main_query,
     get_array_of_selected_fields,
     get_child_court_ids_for_parents,
+    map_to_docket_entry_sorting,
 )
 from cl.search.constants import (
     BOOSTS,
@@ -724,16 +725,6 @@ def add_filter_queries(main_params: SearchParam, cd) -> None:
             main_params["fq"].extend(main_fq)
         else:
             main_params["fq"] = main_fq
-
-
-def map_to_docket_entry_sorting(sort_string: str) -> str:
-    """Convert a RECAP sorting param to a docket entry sorting parameter."""
-    if sort_string == "dateFiled asc":
-        return "entry_date_filed asc"
-    elif sort_string == "dateFiled desc":
-        return "entry_date_filed desc"
-    else:
-        return sort_string
 
 
 def add_grouping(main_params: SearchParam, cd: CleanData, group: bool) -> None:
